+++ packages/e2e.sandbox.officeassignment/src/generatedNoCheck2/index.ts
@@ -1,0 +1,10 @@
+export { endAssignment, recordStatusUpdate, toggleStatusExclusion, updateAssignment } from './ontology/actions.js';
+export * as $Actions from './ontology/actions.js';
+export {} from './ontology/interfaces.js';
+export * as $Interfaces from './ontology/interfaces.js';
+export { Assignment, Employee, Floor, Manager, Office, StatusUpdate } from './ontology/objects.js';
+export * as $Objects from './ontology/objects.js';
+export {} from './ontology/queries.js';
+export * as $Queries from './ontology/queries.js';
+export { $osdkMetadata } from './OntologyMetadata.js';
+export { $ontologyRid } from './OntologyMetadata.js';