+++ packages/maker/src/api/defineAction.ts
@@ -642,7 +642,7 @@ function referencedParameterIds(
   return parameterIds;
 }
 
-function extractAllowedValuesFromActionParameterType(
+export function extractAllowedValuesFromActionParameterType(
   type: ActionParameterType,
 ): ActionParameterAllowedValues {
   if (typeof type === "object") {
@@ -650,6 +650,8 @@ function extractAllowedValuesFromActionParameterType(
       case "objectReference":
       case "objectReferenceList":
         return { type: "objectQuery" };
+      case "objectSetRid":
+        return { type: "objectSetRid" };
       case "struct":
       case "structList":
         throw new Error("Structs are not supported yet");