+++ packages/client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/client
 
+## 2.33.0
+
+### Patch Changes
+
+- @osdk/api@2.33.0
+- @osdk/client.unstable@2.33.0
+- @osdk/generator-converters@2.33.0
+
 ## 2.32.0
 
 ### Minor Changes