+++ packages/widget.client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/widget.client
 
+## 3.30.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.30.0
+
 ## 3.29.0
 
 ### Patch Changes